Fix :set background=dark
When `background` is set to `light` in config, and I set it to `dark` *after* starting Neovim, with `:set background=dark`, I get an error. The issue seems to be with the color `libadwaita_dark`, only defined in the `colors` table when `background=dark`. I think calling `adwaita_dark.set()` from `colors/adwaita.lua` doesn’t regenerate the `colors` table, so the value `libadwaita_dark` is `nil`.

This change seems to regenerate the table. I don’t know if this is the best way to do it, but it seems to fix the issue.